requirements.txt
falcon==2.0.0
peewee==3.14.4
psycopg2==2.8.6
gunicorn==20.1.0
stomp.py==6.1.1
PyJWT==2.0.1
requests==2.25.1
falcon-multipart==0.2.0
paytmchecksum==1.7.0
pycryptodome==3.10.1
paho-mqtt==1.5.1
Pillow==8.2.0
pytz==2018.3
python-dateutil==2.8.2
python-json-logger==0.1.11
# -r requirements-dev.txt # to include dev requirement file